Changed CartItem quantity property type
- Changed cart item quantity property from Integer to Short type.

+++ src/main/java/com/abranlezama/ecommercestore/controller/CartController.java
@@ -37,7 +37,7 @@ public class CartController {
     @PatchMapping
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void updateCartProduct(@Positive @RequestParam("productId") Long productId,
-                                  @Positive @RequestParam("quantity") Integer quantity,
+                                  @Positive @RequestParam("quantity") Short quantity,
                                   Authentication authentication) {
         cartService.updateCartProduct(authentication.getName(), productId, quantity);
     }
