+++ src/main/java/com/abranlezama/ecommercestore/service/CartService.java
@@ -6,9 +6,9 @@ public interface CartService {
 
     CartDTO getCustomerCart(String userEmail);
 
-    void addProductToCart(String userEmail, long productId, int quantity);
+    void addProductToCart(String userEmail, long productId, short quantity);
 
-    void updateCartProduct(String userEmail, long productId, int quantity);
+    void updateCartProduct(String userEmail, long productId, short quantity);
 
     void removeCartProduct(String userEmail, long productId);
 }
